feat(coupon): 重构优惠券模块以符合DDD架构标准
- 引入Contract接口(CouponInput, CouponUserInput)实现依赖抽象
- 实现DTO类(CouponDto, CouponUserDto)用于层间数据传输
- 重构实体类构造函数为无参数形式并添加create()/update()方法
- 增强Mapper类添加getNewEntity()方法统一实体创建入口
- 更新领域服务接受Contract接口参数提升代码可测试性
- 扩展请求层添加toDto()方法实现请求到DTO的转换
- 定义设计文档规范DDD重构的数据流和组件职责划分

+++ app/Domain/Coupon/Contract/CouponInput.php
@@ -23,9 +23,9 @@ interface CouponInput
 
     public function getType(): ?string;
 
-    public function getValue(): ?float;
+    public function getValue(): ?int;
 
-    public function getMinAmount(): ?float;
+    public function getMinAmount(): ?int;
 
     public function getTotalQuantity(): ?int;
